+++ packages/custom-footer-component/CHANGELOG.md
@@ -4,6 +4,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.2](https://github.com/ghiscoding/slickgrid-universal/compare/v3.3.1...v3.3.2) (2023-10-06)
+
+**Note:** Version bump only for package @slickgrid-universal/custom-footer-component
+
+# [3.3.0](https://github.com/ghiscoding/slickgrid-universal/compare/v3.2.2...v3.3.0) (2023-10-05)
+
+**Note:** Version bump only for package @slickgrid-universal/custom-footer-component
+
+## [3.2.2](https://github.com/ghiscoding/slickgrid-universal/compare/v3.2.1...v3.2.2) (2023-09-24)
+
+**Note:** Version bump only for package @slickgrid-universal/custom-footer-component
+
 ## [3.2.1](https://github.com/ghiscoding/slickgrid-universal/compare/v3.2.0...v3.2.1) (2023-09-05)
 
 **Note:** Version bump only for package @slickgrid-universal/custom-footer-component